refactor: api get 커스텀훅으로 분리
